Introduce notion of PLATFORM to Makefile, define simulation platform
