+++ platform/simulation/src/simulation-throw.c
@@ -1,0 +1,8 @@
+#include <ssm-internal.h>
+#include <stdio.h>
+
+void ssm_throw(enum ssm_error reason, const char *file, int line,
+               const char *func) {
+  printf("SSM error at %s:%s:%d: reason: %d\n", file, func, line, reason);
+  exit(1);
+}
